:recycle: Refator Ingame
인게임 관련 코드 리팩토링중+ forceSubmit이 false로 바뀌지 않던 오류 숮ㅇ
