+++ src/components/gameboard/SubmittedPlayer.jsx
