+++ src/pages/GameRoom.jsx
@@ -2,11 +2,9 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import SockJS from 'sockjs-client';
-import * as Stomp from '@stomp/stompjs';
 import AttendeeList from '../components/room/AttendeeList';
 import Button from '../components/common/Button';
-import { setStomp, setIngameHost, closeStomp } from '../app/slices/ingameSlice';
+import { setIngameHost } from '../app/slices/ingameSlice';
 import Container from '../components/layout/Container';
 import FloatBox from '../components/layout/FloatBox';
 import SettingButton from '../components/button/SettingButton';
@@ -177,92 +175,6 @@ function GameRoom() {
     subArray = [];
     token = getCookie(member === 'guest' ? 'guest' : 'access_token');
   }, []);
-
-  useEffect(() => {
-    const client = new Stomp.Client({
-      debug: (str) => {},
-      splitLargeFrames: true,
-      webSocketFactory: () => new SockJS(`${process.env.REACT_APP_API_URL}/ws`),
-    });
-
-    client.onStompError = (frame) => {
-      console.error('Stomp Error!: ', frame.headers.message);
-      console.error('Additional details: ', frame.body);
-    };
-
-    client.onDisconnect = (frame) => {
-      dispatch(closeStomp());
-    };
-    dispatch(setStomp(client));
-
-    const p = new Promise((resolve, reject) => {
-      client.onConnect = (frame) => {
-        subArray.push(
-          client.subscribe(`/topic/game/start/${id}`, (message) => {
-            const data = JSON.parse(message.body);
-            setIsIngame(data.isIngame);
-          }),
-        );
-        subArray.push(
-          client.subscribe(`/topic/game/difficulty/${id}`, (message) => {
-            const data = JSON.parse(message.body);
-            setDifficulty(data.difficulty);
-          }),
-        );
-        subArray.push(
-          client.subscribe(`/topic/game/time-limit/${id}`, (message) => {
-            const data = JSON.parse(message.body);
-            setTimeLimit(data.timeLimit);
-          }),
-        );
-        subArray.push(
-          client.subscribe(`/queue/game/gameroom-data/${socketID}`, (message) => {
-            const data = JSON.parse(message.body);
-            setDifficulty(data.difficulty);
-            setTimeLimit(data.timeLimit);
-          }),
-        );
-
-        client.publish({
-          destination: '/app/game/gameroom-data',
-          body: JSON.stringify({ roomId: id, token, webSessionId: socketID }),
-        });
-
-        resolve();
-      };
-    }).then(() => {
-      dispatch(setStomp(client));
-    });
-
-    return () => {
-      for (let i = 0; i < subArray.length; i += 1) subArray[i].unsubscribe();
-    };
-  }, []);
-
-  useEffect(() => {
-    if (attendees.length > muteUser.length) {
-      const newMuteUser = [];
-      for (let i = muteUser.length; i < attendees.length; i += 1) {
-        newMuteUser.push({
-          nickname: attendees[i].nickname,
-          socketID: attendees[i].socketId,
-          isMuted: false,
-        });
-      }
-      dispatch(setMuteUsers([...muteUser, ...newMuteUser]));
-    } else {
-      const newMuteUsers = [];
-      for (let i = 0; i < attendees.length; i += 1) {
-        for (let j = 0; j < muteUser.length; j += 1) {
-          if (attendees[i].socketId === muteUser[j].socketID) {
-            newMuteUsers.push(muteUser[j]);
-            break;
-          }
-        }
-      }
-      dispatch(setMuteUsers([...newMuteUsers]));
-    }
-  }, [attendees]);
 
   function sendDifficulty(difficulty) {
     ingameStompClient.publish({
